[Prasanna] Code cleanup. Moved interceptors to seperate classes.
