+++ lib/resources/api_base_helper.dart
@@ -4,8 +4,9 @@ import 'dart:convert';
 import 'dart:io';
 import 'package:http/http.dart';
 import 'package:wastexchange_mobile/models/api_exception.dart';
+import 'package:wastexchange_mobile/resources/http_interceptors/auth_interceptor.dart';
+import 'package:wastexchange_mobile/resources/http_interceptors/log_interceptor.dart';
 import 'package:http_interceptor/http_interceptor.dart';
-import 'package:wastexchange_mobile/resources/auth_manager.dart';
 
 class ApiBaseHelper {
   ApiBaseHelper([HttpWithInterceptor http]) {
@@ -68,58 +69,3 @@ class ApiBaseHelper {
   }
 }
 
-/*
-Interceptor that modify API Request by adding authentication information to the request.
- */
-class AuthInterceptor implements InterceptorContract {
-
-  @override
-  Future<RequestData> interceptRequest({RequestData data}) async {
-    await addAuthHeader(data);
-    return data;
-  }
-
-  @override
-  Future<ResponseData> interceptResponse({ResponseData data}) async {
-    return data;
-  }
-
-  Future<void> addAuthHeader(RequestData data) async {
-    final String token = await AuthManager().getAccessToken();
-    if(token != null) {
-      data.headers['Authorization'] = token;
-    }
-  }
-}
-
-///
-/// Interceptor for debug purpose.
-///
-class LogInterceptor implements InterceptorContract {
-
-  @override
-  Future<RequestData> interceptRequest({RequestData data}) async {
-    printRequestLog(data);
-    return data;
-  }
-
-  @override
-  Future<ResponseData> interceptResponse({ResponseData data}) async {
-    printResponseLog(data);
-    return data;
-  }
-
-  void printRequestLog(RequestData data) {
-    print('Method: ${data.method}');
-    print('Url: ${data.url}');
-    print('Body: ${data.body}');
-  }
-
-  void printResponseLog(ResponseData data) {
-    print('Status Code: ${data.statusCode}');
-    print('Method: ${data.method}');
-    print('Url: ${data.url}');
-    print('Body: ${data.body}');
-    print('Headers: ${data.headers}');
-  }
-}
