+++ lib/resources/auth_manager.dart
@@ -5,15 +5,15 @@ import 'package:flutter_secure_storage/flutter_secure_storage.dart';
  */
 class AuthManager {
 
-  static const _accessTokenKey = "access_token";
-
-  AuthManager._internal();
-
-  static final AuthManager _instance = new AuthManager._internal();
-
   factory AuthManager() {
     return _instance;
   }
+
+  AuthManager._internal();
+
+  static final AuthManager _instance = AuthManager._internal();
+
+  static const _accessTokenKey = 'access_token';
 
   void setAccessToken(accessToken) {
     FlutterSecureStorage().write(key: _accessTokenKey, value: accessToken);
@@ -22,5 +22,4 @@ class AuthManager {
   Future<String> getAccessToken() async {
    return FlutterSecureStorage().read(key: _accessTokenKey);
   }
-
 }
