+++ lib/resources/http_interceptors/auth_interceptor.dart
@@ -1,0 +1,24 @@
+import 'package:http_interceptor/http_interceptor.dart';
+import 'package:wastexchange_mobile/resources/auth_manager.dart';
+
+/// Interceptor that modify API Request by adding authentication information to the request.
+class AuthInterceptor implements InterceptorContract {
+
+  @override
+  Future<RequestData> interceptRequest({RequestData data}) async {
+    await addAuthHeader(data);
+    return data;
+  }
+
+  @override
+  Future<ResponseData> interceptResponse({ResponseData data}) async {
+    return data;
+  }
+
+  Future<void> addAuthHeader(RequestData data) async {
+    final String token = await AuthManager().getAccessToken();
+    if(token != null) {
+      data.headers['Authorization'] = token;
+    }
+  }
+}
